requirements-dev.txt
flake8
ipython
ipdb
pytest
pytest-flask
setuptools
flask-shell-ipython